npm run build 実行時に（必要なら）wasm ビルドも走るようにする

+++ build-wasm.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 
+if [ -e src/lyra_wasm.js ]
+then
+  exit 0
+fi
+
 set -eux
 
 cd wasm/
 bazel build -c opt --features=-wasm_warnings_as_errors :lyra-wasm
 cd ../
 
-rm -f src/lyra_wasm.js
 cp wasm/bazel-bin/lyra-wasm/lyra.js src/lyra_wasm.js
